refactor: Refactored cmd line parser
