+++ src/main/java/cl/cavallinux/jisocreator/instances/CommandLineParserManager.java
@@ -2,8 +2,8 @@ package cl.cavallinux.jisocreator.instances;
 
 import org.apache.commons.cli.DefaultParser;
 
-import cl.cavallinux.jisocreator.util.JISOCreatorCommandLineParser;
-import cl.cavallinux.jisocreator.util.cmdline.ICommandLineParser;
+import cl.cavallinux.jisocreator.model.cmdline.ICommandLineParser;
+import cl.cavallinux.jisocreator.model.cmdline.JISOCreatorCommandLineParser;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
