+++ src/main/java/cl/cavallinux/jisocreator/model/cmdline/ICommandLineParser.java
@@ -1,4 +1,4 @@
-package cl.cavallinux.jisocreator.util.cmdline;
+package cl.cavallinux.jisocreator.model.cmdline;
 
 import java.io.IOException;
